[zen-52] Renamed abstract classes in Zend\Mvc
