+++ documentation/manual/it/module_specs/Zend_Mvc-Controllers.xml
@@ -201,7 +201,7 @@ public function plugin($plugin, array $options = null)
   </section>
 
   <section xml:id="zend.mvc.controllers.action-controller">
-    <info><title>L'ActionController</title></info>
+    <info><title>L'AbstractActionController</title></info>
 
     <para>
         Implementare ognuna delle interfacce precedenti è molto
@@ -211,7 +211,7 @@ public function plugin($plugin, array $options = null)
     </para>
 
     <para>
-        Il primo è <classname>Zend\Mvc\Controller\ActionController</classname>.
+        Il primo è <classname>Zend\Mvc\Controller\AbstractActionController</classname>.
         Questo controller implementa ognuna delle interfacce descritte
         precedentemente e usa le seguenti assuzioni:
     </para>
@@ -250,7 +250,7 @@ public function plugin($plugin, array $options = null)
 
     <para>
         Essenzialmente, una mappatura della strada verso
-        un <classname>ActionController</classname> necessita di
+        un <classname>AbstractActionController</classname> necessita di
         ritornare le chiavi "controller" e "action" che combaciano.
     </para>
 
@@ -261,9 +261,9 @@ public function plugin($plugin, array $options = null)
     <programlisting language="php"><![CDATA[
 namespace Foo\Controller;
 
-use Zend\Mvc\Controller\ActionController;
+use Zend\Mvc\Controller\AbstractActionController;
 
-class BarController extends ActionController
+class BarController extends AbstractActionController
 {
     public function bazAction()
     {
@@ -281,7 +281,7 @@ class BarController extends ActionController
       <info><title>Interfacce e collaboratori</title></info>
 
       <para>
-        <classname>ActionController</classname> implementa ognuna di
+        <classname>AbstractActionController</classname> implementa ognuna di
         queste interfacce:
       </para>
 
@@ -346,7 +346,7 @@ class BarController extends ActionController
 
         <listitem>
           <para>
-            <classname>Zend\Mvc\Controller\ActionController</classname>
+            <classname>Zend\Mvc\Controller\AbstractActionController</classname>
           </para>
         </listitem>
       </itemizedlist>
@@ -359,11 +359,11 @@ class BarController extends ActionController
   </section>
 
   <section xml:id="zend.mvc.controllers.restful-controller">
-    <info><title>Il RestfulController</title></info>
+    <info><title>Il AbstractRestfulController</title></info>
 
     <para>
         Il secondo controller astratto che è fornito da ZF2 è
-        <classname>Zend\Mvc\Controller\RestfulController</classname>.
+        <classname>Zend\Mvc\Controller\AbstractRestfulController</classname>.
         Questo controller fornisce una implementazione RESTful che
         semplicemente mappa le richieste HTTP nei metodi del controller
         utilizzando la seguente matrice:
@@ -420,8 +420,8 @@ class BarController extends ActionController
 
     <para>
         In aggiunta, puoi mappare metodi "action" nel
-        <classname>RestfulController</classname>, proprio come si farebbe
-        nell'<classname>ActionController</classname>; questi metodi avranno 
+        <classname>AbstractRestfulController</classname>, proprio come si farebbe
+        nell'<classname>AbstractActionController</classname>; questi metodi avranno
         il suffisso "Action", differenziandosi da metodi RESTful descritti
         precendentemente. Questo ti permette di realizzare azioni come
         per esempio l'invio di form verso diversi metodi RESTful o aggiungere
@@ -432,7 +432,7 @@ class BarController extends ActionController
       <info><title>Interfacce e collaboratori</title></info>
 
       <para>
-        <classname>RestfulController</classname> implementa ognuna delle
+        <classname>AbstractRestfulController</classname> implementa ognuna delle
         seguenti interfacce:
       </para>
 
@@ -498,7 +498,7 @@ class BarController extends ActionController
 
         <listitem>
           <para>
-            <classname>Zend\Mvc\Controller\RestfulController</classname>
+            <classname>Zend\Mvc\Controller\AbstractRestfulController</classname>
           </para>
         </listitem>
       </itemizedlist>
